비지니스로직 및 Usecase 개발

+++ src/main/kotlin/com/yuiyeong/ticketing/presentation/dto/request/ChargingWalletRequest.kt
@@ -1,5 +1,5 @@
 package com.yuiyeong.ticketing.presentation.dto.request
 
 data class ChargingWalletRequest(
-    val amount: Int,
+    val amount: Long,
 )
